+++ pkg/sleepit/sleepit_unix_test.go
@@ -23,7 +23,7 @@ func TestSignalSentToProcessGroup(t *testing.T) {
 		wantOut  []string
 	}{
 		"send 1 signal": {
-			args:     []string{"handle", "-sleep=10ms", "-cleanup=10ms"},
+			args:     []string{"handle", "--sleep=10ms", "--cleanup=10ms"},
 			sendSigs: 1,
 			wantOut: []string{
 				"sleepit: ready\n",
@@ -34,7 +34,7 @@ func TestSignalSentToProcessGroup(t *testing.T) {
 				"sleepit: cleanup done\n"},
 		},
 		"send 2 signals": {
-			args:     []string{"handle", "-sleep=10ms", "-cleanup=10ms"},
+			args:     []string{"handle", "--sleep=10ms", "--cleanup=10ms"},
 			sendSigs: 2,
 			wantOut: []string{
 				"sleepit: ready\n",
@@ -72,7 +72,7 @@ func TestSignalSentToProcessGroup(t *testing.T) {
 			}
 
 			// After the child is started, we want to avoid a race condition where we send
-			// it a signal before it had time to setup its own signal handlers. Sleeping
+			// it a signal before it had time to set up its own signal handlers. Sleeping
 			// is way too flaky, instead we parse the child output until we get a line
 			// that we know is printed after the signal handlers are installed...
 			ready := false
@@ -83,7 +83,7 @@ func TestSignalSentToProcessGroup(t *testing.T) {
 					ready = true
 					break
 				}
-				time.Sleep(10 * time.Millisecond)
+				time.Sleep(time.Millisecond)
 			}
 			if !ready {
 				t.Fatalf(