+++ CHANGELOG.md
@@ -5,6 +5,37 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED
+
+### Fixed
+
+- The displayed duration of each ticker is now correctly rounded to a unit proportional to the value of the `--ticker` flag.
+
+### Breaking
+
+- Due to the introduction of package kong to parse the command-line, now `sleepit` wants flags specified with two hyphens, for example `--sleep` instead of `-sleep`.
+
+### Changed
+
+- Timeit now always reports the command status. For example:
+   ```
+   $ timeit true
+   timeit results:
+       command succeeded
+       real: 4ms
+
+   $ timeit false
+   timeit results:
+       command failed: exit status 1
+       real: 2ms
+
+   $ timeit sleep 2
+   ^Ctimeit: got signal name=interrupt count=1 disposition=ignore
+   timeit results:
+       command terminated abnormally: signal: interrupt
+       real: 1.851s
+   ```
+
 ## [v0.7.0] - [2023-01-21]
 
 ### Fixed
@@ -14,7 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Breaking
 
-- due to the introduction of package kong to parse the command-line, now `timeit` wants flags specified with two hyphens, for example `--ticker` instead of `-ticker`. 
+- Due to the introduction of package kong to parse the command-line, now `timeit` wants flags specified with two hyphens, for example `--ticker` instead of `-ticker`. 
 
 ### Changed